LayoutTests/fast/events/mouse-events-on-textarea-resize-expected.txt


Verifies that correct mouse events are fired and when resizing an element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- test with preventDefault on 'mousedown' ---
--- move mouse into target ---
--- start resizing ---
--- mouse released ---
Received mouseup
Received click

--- test with preventDefault on '' ---
--- move mouse into target ---
--- start resizing ---
Received mousedown
Received focus
--- mouse released ---
Received mouseup
Received click

PASS successfullyParsed is true

TEST COMPLETE
